refactor: 코드리뷰 적용(네이밍 구체화, 유효성 검사 mode 수정 등)
